refactor: 중간병합

+++ src/main/java/org/example/pdnight/global/filter/JwtFilter.java
@@ -103,7 +103,7 @@ public class JwtFilter implements Filter {
                     (String) userNickname,
                     "",
                     userRole,
-                    (Long) userAge,
+                    userAge,
                     userGender,
                     userJobCategory);
